Moved components folder to src instead of src/app
