+++ src/app/page.tsx
@@ -1,11 +1,9 @@
-import HeroSection from "./components/HeroSection";
-import FooterSection from "./components/FooterSection";
+import HeroSection from "../components/HeroSection";
 
 export default function Home() {
   return (
     <main data-test="home-page">
       <HeroSection />
-      <FooterSection />
     </main>
   );
 }
